restructure quizes
