+++ kool/contrib/quiz/question.py
@@ -1,22 +1,14 @@
-"""
-Question
-
-This class represents questions that make up a quiz
-"""
 from kool.db.models import Model
 
 
 class Question(Model):
 	
-	_questions = []
-
-	def __init__(self, question, quiz):
+	def __init__(self, question, quiz_id):
 		super().__init__()
 		self.question = question
-		self.quiz = quiz
+		self.quiz = quiz_id
 		self.choices = {}
 		self.answer = {}
-		_questions.append(self)
 
 	def questions(self):
 		return self._questions
@@ -27,4 +19,13 @@ class Question(Model):
 
 	def set_choices(self, choice):
 		pass
-		
+
+
+class Answer(Model):
+
+	_answers = []
+
+	def __init__(self, student, answer):
+		super().__init__()
+		self.student = student
+		self.answer = answer
