+++ kool/contrib/quiz/quiz.py
@@ -1,16 +1,16 @@
-"""
-Quiz
-
-A base quiz class, providing some default behaviours that all 
-quiz types can inherit or override, as necessary.
-"""
 from kool.db.models import Model
 
 
 class Quiz(Model):
+    """
+    Quiz class
+    """
 
-    def __init__(self, * args, ** kwargs):
+    def __init__(self, name, course_id, **options):
         super().__init__()
-        self.name = kwargs['name']
-        self.description = kwargs['description']
-        self.course = kwargs['course']
+        self.name = name
+        self.course_id = course_id
+        self.description = options.pop('description', None)
+
+
+
